Optimize filtering a bit.
