+++ js/js_bench_test.go
@@ -12,6 +12,9 @@ func BenchmarkCheckEval(b *testing.B) {
 	check := types.FixtureCheck("foo")
 	for i := 0; i < b.N; i++ {
 		synth := dynamic.Synthesize(check)
-		_, _ = js.Evaluate("status == 0", synth, nil)
+		params := map[string]interface{}{
+			"check": synth,
+		}
+		_, _ = js.Evaluate("check.status == 0", params, nil)
 	}
 }
